Move fee multiplier tests into integration_tests
